+++ .storybook/main.ts
@@ -6,8 +6,8 @@ import type { StorybookConfig } from '@storybook/react-vite';
 const config: StorybookConfig = {
   stories: [
     '../*.mdx',
-    '../packages/frontend-tailwind/**/*.mdx',
-    '../packages/frontend-ui-react/**/*.mdx',
+    '../packages/frontend-tailwind/*.mdx',
+    '../packages/frontend-ui-react/src/**/*.mdx',
     '../packages/frontend-ui-react/src/**/*.stories.tsx',
   ],
   staticDirs: ['./public'],