+++ packages/frontend-ui-react/src/index.ts
@@ -4,11 +4,12 @@ export * from './components/pagination';
 export * from './components/localeProvider';
 export * from './components/dialog';
 export * from './components/dialogProvider';
+export * from './components/confirmDialog';
 export * from './components/input';
 export * from './components/clientOnly';
-export * from './components/serverOnly';
 export * from './components/popover';
 export * from './components/calendar';
+export * from './components/alert';
 
 // hooks
 export * from './utils/use-hydrated';