+++ packages/frontend-ui-react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+## 1.4.0
+
+### Minor Changes
+
+- a28cd65: Add Alert component
+
+## 1.3.0
+
+### Minor Changes
+
+- 09c2e5a: add search input (with icon)
+- c0302db: refactor useDialog, so we don't need to call wait()
+- c0302db: add Confirm dialog component/hook
+
+### Patch Changes
+
+- 09c2e5a: fix: input search styling in storybook fails
+- 09c2e5a: fix: button fullWith styling in storybook fails
+
 ## 1.2.0
 
 ### Minor Changes